Some Updates
A vine that drops a rock when shield pulsed (switch on bridge is not
there, had trouble with getting your switch to work)

Fruits should fall when hit by enemies and bullets. They should also
stop on the tar pit to allow you to step on them.

Death Animation for player is implemented.

Falling Platform script is implemented

Lots of playing around with Eli's level. A lot of stuff is not scaled
properly and will need more fixing due to the way it was built.