Wizard Prefab and Character Select
The Wizard prefab is made, he only has 2 animations currently and the
rest are filled with the Witch animations until the rest are turned in.

Menu LOGIC is done. It is not prettied up yet but the important bits and
ideas are there. You click Wizard or Witch and it'll activate one of the
player objects that will need to be in each scene.