Wizard Death updated
Wizard death should play the wizard.

As for the issue of not being able to play the wizard in editor, it is
because in the Menu Manager script there is a private variable that
controls which character you play depending on which one you choose in
the menu. By default it is set to "1" for witch.